냅색응용.cpp 위치변경
